% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrema.R
\name{extrema}
\alias{extrema}
\title{Create Local Extrema Object}
\usage{
extrema(x, interval, min_per_change = NA)
}
\arguments{
\item{x}{A numeric vector or integer vector for use in determining local extrema.}

\item{interval}{An integer value to specify the degree to which changes in local extrema should be
calculated and identified.}

\item{min_per_change}{A numeric value denoting the minimum change to recognize a reversal. Default value is
NA.  NA treated as there is no minimum for reversal.}
}
\description{
This function returns an S3 class, 'local.extrema', which contains local maxima and minima, the
interval used to derive those extrema, and a data frame with xy coordinates that may be used for plotting.
}
\examples{
rv <- extrema(slv, 3)
rv$interval
rv$maxima
rv$minima
}
\references{
base
}
\author{
mjfii
}
\keyword{extrema}
\keyword{maxima}
\keyword{minima}
