% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{local.extrema}
\alias{local.extrema}
\alias{local.extrema-package}
\title{Local Extrema : ...}
\description{
Local Extrema
}
\section{Create Local Extrema Object}{
NA
}

\section{Plot Local Extrema Object}{
NA
}

\section{Demo Data}{
NA
}

