% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extrema.R
\name{plot_extrema}
\alias{plot_extrema}
\title{Plot Local Extrema Object}
\usage{
plot_extrema(x)
}
\arguments{
\item{x}{The only parameter is a `local.extrema` S3 class object.}
}
\description{
This function returns a \code{\link[ggplot2]{ggplot}} visualization based on the results of the \code{\link{extrema}} function.
}
\note{
The function validates that the input is the appropriate class; if it is not, the function will return NULL.
}
\examples{
rv <- extrema(slv, 5)
plot_extrema(rv)
}
\references{
base, ggplot2
}
\author{
mjfii
}
\keyword{extrema}
\keyword{maxima}
\keyword{minima}
